refactor: メソッド切り出しとモジュール化
